Adjust test to match PypeIt repo
Adjust the ``pypeitsetup`` test to match the change in the main PypeIt
repo under the branch of same name.

The ``pypeitsetup.PypeItSetup.build_fitstbl()`` method no longer has a
return value (since this is the only place that used the return value).
The desired table needed for checking is an attribute of the
instantiated object, and this commit completes the same check.

	modified:   unit_tests/test_pypeitsetup.py