Revert Gemini/GMOS modulo fixes to mjd and binning
Revert the Gemini/GMOS ``.pypeit`` files to the OLD state EXCEPT:

There was a bug in the spectrograph class in the past that read/wrote
the MJD as the decimal julian year.  Also, the binning order was
reversed.  This commit reverts the files, but includes the correct
binning and MJD for all files, since the code now uses the values
from the ``.pypeit`` file instead of the headers for configuration-
specific things, and GMOS have date-specific and binning-specific
configurations.

	modified:   pypeit_files/gemini_gmos_gn_e2v_multi_r400_600.pypeit
	modified:   pypeit_files/gemini_gmos_gn_ham_ns_b600_620.pypeit
	modified:   pypeit_files/gemini_gmos_gn_ham_r400_885.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_b480_550.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_b600_mos.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_multi_r400_700.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_r150_869.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_r400_700.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_r400_860.pypeit
	modified:   pypeit_files/gemini_gmos_gs_ham_r400_tilted.pypeit